Jackpot.Board (Jackpot v0.1.0)

The Jackpot board of 9 cells that can be flipped.

Functions

all_boards()

@spec all_boards() :: Jackpot.T.boards()

Generate all 512 possible boards from start (index 0) to winning (index 511).

cell?(board, c)

@spec cell?(Jackpot.T.board(), Jackpot.T.cell()) :: boolean()

Find if the cell is flipped up (true) or not (false).

flip(board, c)

@spec flip(Jackpot.T.board(), Jackpot.T.cell()) :: Jackpot.T.board()

Make a move by flipping a cell from false to true.

It is an error if the cell has already been flipped.

index(board)

@spec index(Jackpot.T.board()) :: Jackpot.T.index()

Calculate the index for a board, by interpreting the cells as binary digits, starting at cell 1 (least significant) to cell 9 (most significant).

move(board, cells, player)

@spec move(Jackpot.T.board(), Jackpot.T.cells(), Jackpot.T.player()) ::
  {:no_move | Jackpot.T.cell(), Jackpot.T.board()}

Let a player make a move for a specific roll in a given position.

The set of possible cell choices has been calculated from the roll. This function wraps the player move by enforcing pre-conditions and checking post-conditions. The cells are filtered to only those allowed by the current position. If there are no allowed cell choices, the game is lost. If there is exactly one cell choice, that move is forced. If there is more than one cell choice, then the player makes the decision. The player's move must be an allowed cell choice.

The function returns :no_move with the losing board position, or the chosen cell move with the new board position.

new()

@spec new() :: Jackpot.T.board()

A new board in the starting position (index 0).

new(index)

@spec new(Jackpot.T.index()) :: Jackpot.T.board()

A board position for a specific index, interpreted as a sequence of binary digits.

win?(board)

@spec win?(Jackpot.T.board()) :: boolean()

Test if a board is the winning position with all cells flipped (index 511).