Jackpot.Graph (Jackpot v0.1.0)

Analyze the play of a player agent across all possible games.

The algorithm uses two passes. The first builds a game graph using board positions as nodes, and player moves as edges between nodes. The second pass traverses the graph from the start position node, propagating probability between layers of the graph, which correspond to successive moves in the games. Finally the win probabillity is read out from the win position node.

Functions

build(player)

@spec build(Jackpot.T.player()) :: Jackpot.T.graph()

Build the game graoh for a player.

Construct all position nodes, add edges according to the player's moves, and initialize the start position probability to 1.0.

calculate_outcomes(player)

@spec calculate_outcomes(Jackpot.T.player()) :: Jackpot.T.statistics()

Get the analytic probability of win/lose for a player.

propagate(graph)

@spec propagate(Jackpot.T.graph()) :: Jackpot.T.graph()

Propagate the starting probability through the graph according to the edge weights between nodes.

The result is a graph where each node has the probability of it being reached in a game. The final node (index 511) has the probability of winning.

render(path)

@spec render(String.t()) :: :ok | {:error, String.t()}

Render the DOT file to PNG, if GraphViz is installed.

statistics(graph)

@spec statistics(Jackpot.T.graph()) :: Jackpot.T.statistics()

Get the winning probability from a game graph.

to_dot(graph, edge_flag, name, dir \\ "output")

@spec to_dot(Jackpot.T.graph(), boolean(), String.t(), String.t()) ::
  {path :: String.t(), dot :: IO.chardata()}

Convert a graph to a GraphViz DOT format file. Write the DOT data to file.

If edge_flag is true then draw the cell move on edges, otherwise do not label edges, but do draw layer boxes and titles.