Jackpot.T (Jackpot v0.1.0)

Types for the Jackpot application.

Types

board()

@type board() :: %{required(cell()) => boolean()}

The board is a row of 9 cells. The cells are held in a map for fast lookup.

boards()

@type boards() :: %{required(index()) => board()}

The set of all possible boards, accessed by integer index.

cell()

@type cell() :: 1..9

A single cell on the board, which can be down (false) or flipped up (true). Cells are ordered and numbered from left (1) to right (9).

cells()

@type cells() :: [cell()]

A set of cells that can be played from a roll of 2 dice. The list calculated from the roll will have 1, 2 or 3 elements.

The set may be filtered by removing cells that are already flipped (true). The result is a set of possible moves in the game, which may be empty if all the cells are already flipped up.

count()

@type count() :: 1..36

A count of the number of rolls of 2 dice that lead to the same cell choice.

die()

@type die() :: 1..6

A single die with numbers 1..6

edges()

@type edges() :: %{required(index()) => count()}

The outgoing edges from a board position node to the next position node in the game Graph. The edge corresponds to the flipping of exactly one cell in the board. The edges are stored as the destination node (board position index) with the number of dice rolls for which the player would choose that move.

For example, the 15 rolls {1,3} {3,1} {2,2} {4,} {,4} might all result in flipping cell 4 and moving to the same next position node.

game()

@type game() :: {board(), player(), moves()}

A game is the current board position, a player, and the historical sequence of moves already played. A game is complete if it is won or lost.

A winning game has a board with all cells flipped (index 511) and 9 moves containing all cells exactly once, without any :no_move blocked position.

A losing game has a board with at least one cell not flipped, and the last move is a blocked position with :no_move.

graph()

@type graph() :: %{required(index()) => {edges(), probability()}}

A game graph is a directed graph containing board positions as nodes, and moves as edges. The set of nodes contains all 512 possible board positions. Each node has up to 36 outgoing edges corresponding to the allowed legal moves chosen by a player agent.

The graph is layered by how many moves have been made (i.e. how many cells are flipped) from 0 (start) to 9 (win). There are no edges within a layer. Each layer only has edges going out to members of the next highest layer. The root (source) node is the starting position (index 0), which does not have any incoming edges. The final (sink) node is the winning position (index 511), which does not have any outgoing edges.

Each node also stores the probability of reaching that position, which is the sum of probabilites along each path of edges leading to the position.

index()

@type index() :: 0..511

The index of a board position is the integer represented by treating cells as binary digits. The bits are shifted by cell number, so the binary number is read from left (1st - least significant) to right (9th - most signiificant). The start position with all cells off (0, false) has index 0. The winning position with all cells on (1, true) has index 511 which is 2^9 - 1.

move()

@type move() :: {roll(), :no_move | cell()}

A move in the game is a roll of 2 dice, with a choice of which cell to flip, or :no_move if there is no valid move and the game is lost.

moves()

@type moves() :: [move()]

A game history is a sequence of moves.

outcome()

@type outcome() :: :win | :lose

The outcome of a game is always win or lose.

percent()

@type percent() :: float()

Probability of winning or losing as a percentage.

player()

@type player() :: module()

A player is implemented as a module with a move function.""

probability()

@type probability() :: float()

Probability of some event in the game.

result()

@type result() :: {outcome(), board(), [move()]}

The result of a game is an outcome, the final board position, and the sequence of moves that were played.

roll()

@type roll() :: {die(), die()}

A roll of 2 dice with numbers 1..6

statistics()

@type statistics() :: %{required(outcome()) => percent()}

Statistics for winning or losing.

Functions

is_cell(c)
(macro)

is_count(n)
(macro)

is_die(i)
(macro)

is_index(ix)
(macro)

is_move(m)
(macro)

is_non_empty_cells(cs)
(macro)

is_pc(p)
(macro)

is_prob(p)
(macro)

is_roll(r)
(macro)