Jackpot.Graph (Jackpot v0.1.0)

Analyze the play of a player agent across all possible games.

The algorithm uses two passes. The first builds a game graph using board positions as nodes, and player moves as edges between nodes. The second pass traverses the graph from the start position node, propagating probability between layers of the graph, which correspond to successive moves in the games. Finally the win probabillity is read out from the win position node.

Functions

calculate_outcomes(player)

@spec calculate_outcomes(Jackpot.T.player()) :: Jackpot.T.statistics()

Get the analytic probability of win/lose for a player.